refactor: split processOrder method
Check for low stock products after order completion, invalidate cache if
any
